Place commands into table 'loveos.commands'
Problem:

The current method for checking if a command is valid, is to look
up if it exists as a global variable.
This means that any valid global variable can be a valid command.
For example, if the user enters 'math' as a command, love will crash with
'attempting to call a table value' error.

Solution:

By placing all commands into the table loveos.commands, only names that
are stored in that table are valid commands.